refactor(storage): extract agents module (R.1) — modularization complete
Final R-phase of the storage modularization plan
(SPEC_STORE_MODULARIZATION_2026_05_27.md). Moves the entire agent
subsystem out of store.rs (4836 → 3306 lines, -1530) into a new
`agents.rs` (1556 LOC):

Structs:
- `AgentDefinition` (the 24-field template/user-clone def row)
- `AgentInstance` (per-launch instance with bindings + lifecycle)
- `InstanceStatus` enum + impl (running/paused/stopped/crashed/detached)

Module-level helpers:
- `derive_slug` (filesystem-safe slug from display name)
- `default_agent_type` / `default_context_json` (serde defaults)
- `map_instance_row` / `map_agent_definition_row` (row mappers shared
  between def + instance queries)

Methods (20 total, across two `impl Store {}` blocks in the new file):
- 8 `agent_def_*` — get / list / count / delete_seeded / insert /
  set_hidden / update / delete
- 12 `instance_*` — list / get / create / set_hidden / list_named /
  get_by_name / update / repoint_definition / delete /
  backfill_identity_id / get_active_for_block / list_legacy
  (the last is a private helper used by Phase 3b.3b fallback)

Method surface is unchanged: `Store::agent_def_*` and
`Store::instance_*` still live on `Store` via the new module's two
`impl Store {}` blocks. All 3 public types re-exported from
`storage::store::*` AND `storage::*` so existing imports stay
green — verified across `server/agent_handlers.rs`,
`backend/agent_session.rs`, and `server/app_api.rs`.

One visibility tweak: `Store::registry()` bumped from `fn` to
`pub(super) fn` so the sibling agents module can take the
registry handle for the `registry_upsert_if_named` calls embedded
in instance lifecycle methods. The public `shared_agent_registry()`
accessor is unchanged.

162/162 storage tests + 9/9 agent_def + 55/55 instance tests green.

## R-phase complete

| Phase | Status |
|---|---|
| R.0 rename | ✅ #1100 |
| R.1 agents.rs | ✅ this PR |
| R.2 identities.rs | ✅ #1127 |
| R.3 memory_bundles | ✅ #1116 |
| R.4 content/skills/history | ✅ #1126 |
| R.5 dual_write | ✅ #1128 |
| R.6 registry_mirror | ✅ #1129 |

store.rs is now 3306 LOC of focused infrastructure (connection mgmt,
generic CRUD, registry accessors, re-exports). Every Rust file in
the repo is under 1,800 LOC except for this PR's agents.rs
(intentionally — the agent subsystem really is that large).
