+++ .changesets/1779990926-refactor-storage-extract-agents-module-r-1-modularization-co-n8q9.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(storage): extract agents module (R.1) — modularization complete
